Move get_target_group_name function to helpers.rb
