+++ cookbooks/aws-parallelcluster-slurm/libraries/helpers.rb
@@ -15,6 +15,9 @@
 #
 # Retrieve compute nodename from file
 #
+
+require 'digest'
+
 def slurm_nodename
   slurm_nodename_file = "#{node['cluster']['slurm_plugin_dir']}/slurm_nodename"
 
@@ -136,4 +139,12 @@ def get_primary_ip
   end
 
   primary_ip
+end
+
+def get_target_group_name(cluster_name, pool_name)
+  partial_cluster_name = cluster_name[0..6]
+  partial_pool_name = pool_name[0..6]
+  combined_name = cluster_name + pool_name
+  hash_value = Digest::SHA256.hexdigest(combined_name)[0..15]
+  "#{partial_cluster_name}-#{partial_pool_name}-#{hash_value}"
 end
