+++ cookbooks/aws-parallelcluster-slurm/recipes/config/config_check_login_stopped_script.rb
@@ -12,28 +12,18 @@
 # limitations under the License.
 # rubocop:disable Style/SingleArgumentDig
 
-require 'digest'
-
-def get_target_group_name(cluster_name, pool_name)
-  partial_cluster_name = cluster_name[0..6]
-  partial_pool_name = pool_name[0..6]
-  combined_name = cluster_name + pool_name
-  hash_value = Digest::SHA256.hexdigest(combined_name)[0..15]
-  "#{partial_cluster_name}-#{partial_pool_name}-#{hash_value}"
-end
-
 template "#{node['cluster']['scripts_dir']}/slurm/check_login_nodes_stopped.sh" do
   source 'slurm/head_node/check_login_nodes_stopped.sh.erb'
   owner 'root'
   group 'root'
   mode '0700'
   variables(
-    target_group_name: lazy {
+    target_group_name: lazy do
       get_target_group_name(
         node['cluster']['cluster_name'] || node['cluster']['stack_name'],
         node['cluster']['config'].dig(:LoginNodes, :Pools, 0, :Name)
       )
-    },
+    end,
     region: node['cluster']['region']
   )
   only_if do
